requirements.txt
streamlit
tensorflow
numpy
pillow
pickle-mixin